Move the initialization part to the separate file Par_Output_TracerParticle_Mesh.cpp
